refactor(typescript): Fix imports
