+++ src/elements/content-sidebar/activity-feed/task-form/TaskError.js
@@ -10,7 +10,7 @@ import getProp from 'lodash/get';
 import messages from './messages';
 import apiMessages from '../../../../api/messages';
 import { TASK_EDIT_MODE_EDIT, TASK_MAX_GROUP_ASSIGNEES, ERROR_CODE_GROUP_EXCEEDS_LIMIT } from '../../../../constants';
-import InlineNotice from '../../../../components/inline-notice/InlineNotice';
+import InlineNotice from '../../../../components/inline-notice';
 
 import type { TaskType, TaskEditMode } from '../../../../common/types/tasks';
